+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+-  add AsterixPacket::finalize() for updating the packet to the correct fspec and len after
+   messages are added
 
 ## [0.1.1] - 2020-08-16
 -  Add License file (MIT) and prepare Cargo.toml file for release